Add search and random search functionalities
wikiSearch() takes the searchText as the argument and performs the
search using MediaWiki REST action API. The result will be parsed and
stored in a global variable called "result" which will be used later
when AJAX is completed.

When user presses ENTER key in the search box or presses search button,
the search text is checked against empty and is passed to wikiSearch().
When the AJAX is completed, the result div will be updated accordingly.

Random search is performed by calling to Special:Random.